+++ Aurora/External/libRocket/Source/Core/ElementTextDefault.cpp
@@ -14,7 +14,7 @@
  *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -81,8 +81,8 @@ void ElementTextDefault::OnRender()
 	FontFaceHandle* font_face_handle = GetFontFaceHandle();
 	if (!font_face_handle)
 		return;
-	
-	
+
+
 	// If our font configuration has potentially changed, update it and force a geometry
 	// generation if necessary.
 	if (font_dirty &&
@@ -96,31 +96,31 @@ void ElementTextDefault::OnRender()
 		GenerateGeometry(font_face_handle);
 
 	Vector2f translation = GetAbsoluteOffset();
-	
+
 	bool render = true;
 	Vector2i clip_origin;
 	Vector2i clip_dimensions;
 	if (GetContext()->GetActiveClipRegion(clip_origin, clip_dimensions))
 	{
-		float clip_top = clip_origin.y;
-		float clip_left = clip_origin.x;
-		float clip_right = (clip_origin.x + clip_dimensions.x);
-		float clip_bottom = (clip_origin.y + clip_dimensions.y);
-		float line_height = GetFontFaceHandle()->GetLineHeight();
-		
+		float clip_top = (float)clip_origin.y;
+		float clip_left = (float)clip_origin.x;
+		float clip_right = (float)(clip_origin.x + clip_dimensions.x);
+		float clip_bottom = (float)(clip_origin.y + clip_dimensions.y);
+		float line_height = (float)GetFontFaceHandle()->GetLineHeight();
+
 		render = false;
 		for (size_t i = 0; i < lines.size(); ++i)
-		{			
+		{
 			const Line& line = lines[i];
 			float x = translation.x + line.position.x;
 			float y = translation.y + line.position.y;
-			
+
 			bool render_line = !(x > clip_right);
 			render_line = render_line && !(x + line.width < clip_left);
-			
+
 			render_line = render_line && !(y - line_height > clip_bottom);
 			render_line = render_line && !(y < clip_top);
-			
+
 			if (render_line)
 			{
 				render = true;
@@ -128,7 +128,7 @@ void ElementTextDefault::OnRender()
 			}
 		}
 	}
-	
+
 	if (render)
 	{
 		for (size_t i = 0; i < geometry.size(); ++i)
@@ -225,7 +225,7 @@ bool ElementTextDefault::GenerateLine(WString& line, int& line_length, float& li
 		// The token can fit on the end of the line, so add it onto the end and increment our width and length
 		// counters.
 		line += token;
-		line_length += (next_token_begin - token_begin);
+		line_length += (int)(next_token_begin - token_begin);
 		line_width += token_width;
 
 		// Break out of the loop if an endline was forced.
